requirements.txt
langchain>=0.2.0
faiss-cpu>=1.7.4
sentence-transformers>=2.7.0
llama-cpp-python>=0.2.67
python-dotenv>=1.0.1
pypdf>=3.19.4
tqdm>=4.66.4
jupyterlab>=4.2.0